+++ src/main/java/org/thoughtslive/jenkins/plugins/jira/service/JiraEndPoints.java
@@ -8,6 +8,7 @@ import retrofit2.http.GET;
 import retrofit2.http.POST;
 import retrofit2.http.PUT;
 import retrofit2.http.Path;
+import retrofit2.http.Query;
 
 /**
  * JIRA Restful Endpoints.
@@ -125,5 +126,12 @@ public interface JiraEndPoints {
 
   @POST("rest/api/2/field")
   Call<Object> createField(@Body Object field);
+  
+  // Users
+  @GET("rest/api/2/user/search")
+  Call<Object> userSearch(@Query("username") String userName, @Query("startAt") int startAt, @Query("maxResults") int maxResults);
+
+  @GET("rest/api/2/user/assignable/search")
+  Call<Object> assignableUserSearch(@Query("username") String userName, @Query("project") String project, @Query("issueKey") String issueKey, @Query("startAt") int startAt, @Query("maxResults") int maxResults);
 
 }